Don't expose initGameObject
