+++ src/index.ts
@@ -1,7 +1,6 @@
 import type { InjectionKey } from 'vue'
 import type { UpdateEventHandler } from './components/Scene.vue'
-import * as Phaser from 'phaser'
-import { customRef, getCurrentInstance, inject, onBeforeUnmount, watch } from 'vue'
+import { customRef, inject, onBeforeUnmount } from 'vue'
 import Body from './components/Body.vue'
 import Circle from './components/Circle.vue'
 import Container from './components/Container.vue'
@@ -36,113 +35,6 @@ import TilemapLayer from './components/TilemapLayer.vue'
 import Triangle from './components/Triangle.vue'
 import Zone from './components/Zone.vue'
 import { InjectionKeys, PrivateInjectionKeys } from './provider'
-import setters, { deepProps, GAME_OBJECT_EVENTS, vModelProps } from './setters'
-
-function createPhavuerApp() {
-  console.error('Phavuer::createPhavuerApp() has been removed. Please use `<Game>` component instead. See: https://github.com/laineus/phavuer')
-}
-
-const camelize = (s: string) => s.replace(/-./g, x => x[1].toUpperCase())
-
-function checkIsBody(object: any): object is Phaser.Physics.Arcade.Body | Phaser.Physics.Arcade.StaticBody {
-  return object && 'bounce' in object
-}
-function checkIsFx(_object: any, isFx: boolean): _object is Phaser.Filters.Controller {
-  return isFx
-}
-function checkIsLight(object: any): object is Phaser.GameObjects.Light {
-  return object.constructor === Phaser.GameObjects.Light
-}
-
-function initGameObject(
-  object: Phaser.GameObjects.GameObject | Phaser.Physics.Arcade.Body | Phaser.Physics.Arcade.StaticBody | Phaser.Filters.Controller | Phaser.GameObjects.Light,
-  props: Readonly<Record<string, unknown>>,
-  options: { isFx?: boolean } = {},
-) {
-  const currentInstance = getCurrentInstance()!
-  const emit = currentInstance.emit.bind(currentInstance)
-  const isLight = checkIsLight(object)
-  const isFx = checkIsFx(object, !!options.isFx)
-  const isBody = checkIsBody(object)
-  const isGameObject = !isLight && !isFx && !isBody
-  const scene = inject(InjectionKeys.Scene)!
-  const renderList = inject(PrivateInjectionKeys.RenderTextureRenderList)
-  if (isLight) {
-    if (!scene.lights.active)
-      scene.lights.enable()
-    scene.lights.lights.push(object)
-  }
-  else if (isBody) {
-    // _
-  }
-  else if (isFx) {
-    // FX objects don't need to be added to the scene
-  }
-  else if (renderList) {
-    renderList.push(object)
-    onBeforeUnmount(() => {
-      const i = renderList.findIndex(v => v === object)
-      renderList.splice(i, 1)
-    })
-  }
-  else {
-    scene.add.existing(object)
-    // Append to parent container
-    const container = inject(InjectionKeys.Container)
-    if (container) {
-      const i = container.list.findIndex((v) => {
-        return ('depth' in v ? (v as Phaser.GameObjects.GameObject & Phaser.GameObjects.Components.Depth).depth : 0) > ((props.depth as number) ?? 0)
-      })
-      i === -1 ? container.add(object) : container.addAt(object, i)
-    }
-  }
-  // Make it reactive
-  const definedProps = Object.fromEntries(
-    Object.entries(currentInstance.vnode.props ?? {}).map(([key, value]) => [camelize(key), value]),
-  )
-  const vModelKeys = Object.keys(definedProps).filter(key => key.startsWith('onUpdate:')).map(key => key.split(':')[1]).filter(key => vModelProps.includes(key))
-  // @ts-expect-error ts(7053)
-  const normalProps = Object.entries(definedProps).filter(([key]) => setters[key])
-  const watchStoppers = normalProps.map(([key, value]) => {
-    // @ts-expect-error ts(7053)
-    const setter = setters[key](object, vModelKeys.includes(key) ? emit : undefined)
-    setter(value)
-    // TODO: Don't watch non dynamicProps
-    return watch(() => props[key], setter, { deep: deepProps.includes(key) })
-  }).filter(Boolean)
-  // Set event
-  if (definedProps.onCreate)
-    emit('create', object)
-  // Set interactive events
-  const events = GAME_OBJECT_EVENTS.filter(v => v.attr in definedProps)
-  if (isGameObject && events.length) {
-    if (!object.input) {
-      object.setInteractive()
-    }
-    if (events.some(v => v.drag)) {
-      scene.input.setDraggable(object)
-    }
-    events.forEach((v) => {
-      object.on(v.emit, (...args: any) => {
-        if ('eventIndex' in v) {
-          args[0].stopPropagation = args[v.eventIndex as number].stopPropagation
-        }
-        emit(v.emit, ...args)
-      })
-    })
-  }
-  // Destroy when unmounted
-  onBeforeUnmount(() => {
-    watchStoppers.forEach(stop => stop())
-  })
-  if (isLight) {
-    onBeforeUnmount(() => scene.lights.removeLight(object))
-  }
-  else {
-    onBeforeUnmount(() => object.destroy())
-  }
-  return object
-}
 
 function useInject<T>(key: InjectionKey<T>) {
   return () => {
@@ -194,7 +86,6 @@ export {
   Body,
   Circle,
   Container,
-  createPhavuerApp,
   FxBarrel,
   FxBloom,
   FxBlur,
@@ -211,7 +102,6 @@ export {
   FxWipe,
   Game,
   Image,
-  initGameObject,
   InjectionKeys,
   Light,
   Line,
@@ -221,7 +111,6 @@ export {
   Polygon,
   Rectangle,
   refPhaserInstance,
-  refTo,
   RenderTexture,
   RoundRectangle,
   Scene,
